api/requirements-to-freeze.txt
tornado
pynsq

# testing
pytest
ipython
ipdb
